pyproject.toml
[tool.poetry]
name = "libb-cmd"
version = "0.1.0"
description = ""
authors = ["bissli <bissli@xyz.com>"]
readme = "README.md"
packages = [{ include = "cmdline", from = "src" }]

[tool.poetry.dependencies]
python = "~3.11"

libb-log = { git = "https://github.com/bissli/libb-log.git", rev = "master" }

# == test
asserts = { version = "*", optional = true }
pytest = { version = "*", optional = true }
pytest-mock = { version = "*", optional = true }
pytest-runner = { version = "*", optional = true }
wrapt = { version = "*", optional = true }
pdbpp = { git = "https://github.com/pdbpp/pdbpp.git", rev = "master", optional = true }

[tool.poetry.extras]
test = [
  "asserts",
  "docker",
  "pdbpp",
  "pytest",
  "pytest-mock",
  "pytest-runner",
  "wrapt",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
